Explore using recursive depth-first search
Implement the explore() phase using a recursive depth-first search. A
depth-first search isn't very efficient in exploring our cavern, which
only has one exit, but it's quite elegant to implement recursively and
it's good starting point to help me understand graph searches.

The `nodesSeen` set is used to remember which nodes we've already
visited, to avoid us looping in circles, and we return to the original
position we started from after recursively calling `explore()` so that
we can backtrack if all neighbouring nodes have already been visited.

The instance variable `nodesSeen` probably has the wrong scope, but for
this first effort it's a convenient way of keeping state when calling
the recursive method. I'd refactor this if I were going to stick with
this approach.