Escape using Dijkstra's algo with priority queue
Implement the escape phase using Dijkstra's shortest path algorithm in
combination with a priority queue:

https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm

The outline of the code borrows heavily from the
`minPathLengthToTarget()` method from the `Cavern` class, which is used
to determine the shortest path to the orb or the exit by the game
itself, for the purposes of determining how much time we should be
allowed:
https://github.com/keithmannock/BBK-PiJ-2015-16/blob/5a3b52b25dd7bfd81378efbe7964b54667c2d04f/cw-temple/src/game/Cavern.java#L393-L431

It made sense to mirror re-use this code here. I've added an additional
data structure, `prevNode`, which is used to record the shortest route
so that we can play it back later when moving George around the temple.