Rename InternalMinHeap to MinHeap; make public
I plan to reuse this data structure when implementing the escape phase
in the `student` package, so make the `InternalMinHeap` class public and
rename the class accordingly.

I did consider whether I should extract this class into its own package,
since it's now quite self-contained, but for now I think the name change
is sufficient. My code in the `student` package already calls a lot of
classes in the `game` package, so I think it's okay to call this one
too.