Explore: switch to breadth-first search
Replace my original depth-first search with a breadth-first search
during the explore phase. This implementation is iterative rather than
recursive.

We use the `nodesSeen` variable to record which nodes have already been
visited, to avoid loops. The `lastVisited` deque is used to backtrack if
we move George into a corner where all the neighbouring tiles have
already been visited.

Anecdotal testing shows that this search method is much more efficient
and the bonus multiplier is usually much closer to the 1.3 maximum than
it was using a depth-first search.

I believe breadth-first is more likely efficient in this use case
because there is only one target (one orb).

Note that this implementation searches for the shortest path only; it
does not make any attempt to alter the path based on the available gold.